gcp: Add most install components
This commit adds all GCP install components except for the org-wide-policy
component. This will come in a later PR.

This work uses the new root metadata on the GCP root install, as well as
item-specific metadata.

Components are structured as follows:

- p0_gcp: The root organization install; holds metadata for the access-logs
  component
- p0_access_logs: Installable without staging, as proof of project ownership
  is determined by having a currently installed iam-assessment component
- p0_iam_assessment: Requires a staging install to prove ownership
- p0_iam_write: Requires a staging install to prove ownership
- p0_sharing_restriction: Installable without staging as it depends on an
  installed iam-write component

For each component, I added detailed instructions on peer Terraform that must
be installed alongside.