Handle resources removed in environment
Previously, application would fail with a generic HTTP error if a
resource was removed in the environment, but kept in the configuration.

Instead, detect deleted resources, so that they can be re-added on
application.

Note that there's an opportunity to DRY the generic-resource logic,
but I've intentionally skipped that for now.